refac(app): reface @/lib dir to be organized by operations, states, types directories.
